More work porting back to headers/cpp's
- First successful run with geometry setup + placeholder draw work; moving draw work across next

- Ported all files needed for geometry setup across from modules, fiddled around with file layouts to minimize errors

- Realized I wasn't freeing global memory on shutdown, added that in

- Fixed up project settings (all configs take SimpleTiling.lib as input & build with C++20)

+++ vox-sculpt-refactor/vox-sculpt-refactor/path.cpp
@@ -14,7 +14,7 @@ void tracing::path::resolve_path_weights(float* rho_out, float* pdf_out, float* 
     float response = 1.0f;
     float power = 1.0f;
     *rho_out = vts[0].rho_sample;
-    for (uint32_t i = 0; i < size; i++)
+    for (int16_t i = 0; i < size; i++)
     {
         auto& bounce = vts[i];
         pdf *= bounce.pdf;
